+++ mathics/builtin/compile/ir.py
@@ -339,7 +339,7 @@ class IRGenerator(object):
             return exponent
 
         # E ^ exponent
-        if leaves[0].same(Symbol('E')) and exponent.type == real_type:
+        if leaves[0].sameQ(Symbol('E')) and exponent.type == real_type:
             return self.call_fp_intr('llvm.exp', [exponent])
 
         # 2 ^ exponent